Remove old timer
Timer originally used an npm library, but this was raising warnings as it
required changing state within render (due to requirement to stop timer using
props).

This custom timer simplifies logic of stopping the count using `isRecording`
props.
