+++ src/components/Timer/timer.tsx
@@ -1,43 +1,41 @@
-import React, { Fragment, Component } from 'react';
-import Timer from 'react-compound-timer';
+import React, { useState, FunctionComponent, useEffect } from 'react';
 import styles from './timer.module.css';
 
-interface TimerProps {
+interface RecordingProps {
   isRecording: boolean;
 }
 
-interface TimerMethods {
-  start: () => void;
-  stop: () => void;
-  reset: () => void;
-}
+export const Timer: FunctionComponent<RecordingProps> = ({isRecording}: RecordingProps) => {
+  const [ seconds, setSeconds ] = useState(0);
+  const [ minutes, setMinutes ] = useState(0);
 
-export class RecordingTimer extends Component<TimerProps> {
-  timerController({ start, stop, reset }: TimerMethods) {
-    if (this.props.isRecording) {
-      start();
-    }
-    if (!this.props.isRecording) {
-      stop();
-    }
-  }
+  const formatValue = (value:number) => `${value < 10 ? `0${value}` : value}`
 
-  render() {
-    return (
-      <div className={styles.timer}>
-        <Timer
-          startImmediately={false}
-          formatValue={(value) => `${value < 10 ? `0${value}` : value}`}
-        >
-          {({ start, stop, reset }: TimerMethods) => (
-            <Fragment>
-              <Timer.Minutes />:
-              <Timer.Seconds />
-              {this.timerController({ start, stop, reset })}
-            </Fragment>
-          )}
-        </Timer>
-      </div>
-    );
-  }
+  useEffect(() => {
+    let interval:any;
+    if (isRecording) {
+      interval = setInterval(() => {
+        setTime()
+      }, 1000);
+    } else if (!isRecording && seconds !== 0) {
+      clearInterval(interval);
+    }
+
+    const setTime = () => {
+      setSeconds(seconds => seconds + 1);
+        if (seconds === 59 ) {
+          setMinutes(minutes => minutes + 1);
+          setSeconds(0)
+        }
+    }
+
+    return () => clearInterval(interval);
+    }, [isRecording, seconds]);
+
+
+  return (
+    <div className={styles.timer}>
+      {formatValue(minutes)}:{formatValue(seconds)}
+    </div>
+  );
 }
